importProductsFile
